Massively simplified config
